filters in other file and logout function
