+++ FrontEnd/index.html
@@ -27,7 +27,8 @@
 			<ul>
 				<li class="activ-page">projets</li>
 				<li><a href="#contact">contact</a></li>
-				<li><a href="./pages/login.html">login</a></li>
+				<li><a id="login-link" href="./pages/login.html">login</a></li>
+				<!-- <li><a id="logout-link" href="#">logout</a></li> -->
 				<li><img src="./assets/icons/instagram.png" alt="Instagram"></li>
 			</ul>
 		</nav>
@@ -37,16 +38,16 @@
 	<section id="introduction">
 		<figure>
 			<img src="./assets/images/sophie-bluel.png" alt="">
-			<div class="edit edit--author-img">
+			<a class="edit edit--author-img">
 				<span class="material-symbols-outlined">edit_square</span>
 				<span>modifier</span>
-			</div>
+			</a>
 		</figure>
 		<article>
-			<div class="edit edit--intro">
+			<a class="edit edit--intro">
 				<span class="material-symbols-outlined">edit_square</span>
 				<span>modifier</span>
-			</div>
+			</a>
 			<h2>Designer d'espace</h2>
 			<p>Je raconte votre histoire, je valorise vos idées. Je vous accompagne de la conception à la livraison finale du chantier.</p>
 			<p>Chaque projet sera étudié en commun, de façon à mettre en valeur les volumes, les matières et les couleurs dans le respect de l’esprit des lieux et le choix adapté des matériaux. Le suivi du chantier sera assuré dans le souci du détail, le respect du planning et du budget.</p>
@@ -56,10 +57,10 @@
 	<section id="portfolio">
 			<div class="portfolio-title">
 				<h2>Mes Projets</h2>
-				<div class="edit edit--portfolio">
+				<a class="edit edit--portfolio">
 					<span class="material-symbols-outlined">edit_square</span>
 					<span>modifier</span>
-				</div>
+				</a>
 			</div>
 			<div class="filters">
 				<!-- here we add the filters depending on the category names available-->
